// webapp/src/app/core/api/v1/model/applicationSection.ts
/**
 * permaAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApplicationSection { 
    applicationSectionId?: number;
    sectionTitle?: string | null;
    description?: string | null;
    isActive?: boolean;
    sortOrder?: number;
    subTitle?: string | null;
    isCompleted?: boolean;
}