// webapp/src/app/core/api/v1/model/elementType.ts
/**
 * permaAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ElementType = 'Currency' | 'LargeText' | 'Checkbox' | 'Text' | 'Integer' | 'Table' | 'SingleSelect' | 'MultiSelect' | 'HTML' | 'Attachment' | 'Email' | 'Date';

export const ElementType = {
    Currency: 'Currency' as ElementType,
    LargeText: 'LargeText' as ElementType,
    Checkbox: 'Checkbox' as ElementType,
    Text: 'Text' as ElementType,
    Integer: 'Integer' as ElementType,
    Table: 'Table' as ElementType,
    SingleSelect: 'SingleSelect' as ElementType,
    MultiSelect: 'MultiSelect' as ElementType,
    Html: 'HTML' as ElementType,
    Attachment: 'Attachment' as ElementType,
    Email: 'Email' as ElementType,
    Date: 'Date' as ElementType
};